fix TooManyFunctions
